+++ src/main/java/uk/co/real_logic/benchmarks/rtt/aeron/LiveReplayNode.java
@@ -33,6 +33,7 @@ import static uk.co.real_logic.benchmarks.rtt.aeron.AeronUtil.*;
 
 /**
  * Remote node which subscribes to the replay channel of the archive and send replayed messages to the sender.
+ * Counterpart for the {@link ArchiveMessageTransceiver}.
  */
 public final class LiveReplayNode implements AutoCloseable
 {
