+++ src/test/java/uk/co/real_logic/benchmarks/rtt/aeron/LiveRecordingTest.java
@@ -17,7 +17,6 @@ package uk.co.real_logic.benchmarks.rtt.aeron;
 
 import io.aeron.archive.ArchivingMediaDriver;
 import io.aeron.archive.client.AeronArchive;
-import org.agrona.CloseHelper;
 import org.agrona.LangUtil;
 import org.agrona.collections.LongArrayList;
 import org.junit.jupiter.api.Test;
@@ -26,37 +25,38 @@ import uk.co.real_logic.benchmarks.rtt.Configuration;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
-import java.util.stream.LongStream;
 
 import static io.aeron.archive.client.AeronArchive.connect;
 import static java.lang.Long.MIN_VALUE;
+import static java.util.stream.LongStream.range;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static uk.co.real_logic.benchmarks.rtt.aeron.AeronUtil.launchArchivingMediaDriver;
 
-class LiveReplayMessageTransceiverTest
+class LiveRecordingTest
 {
     @Test
     void test() throws Exception
     {
+
         final int messages = 1_000_000;
         final Configuration configuration = new Configuration.Builder()
             .numberOfMessages(messages)
-            .messageTransceiverClass(LiveReplayMessageTransceiver.class)
+            .messageTransceiverClass(LiveRecordingMessageTransceiver.class)
             .build();
 
-        final ArchivingMediaDriver archivingMediaDriver = launchArchivingMediaDriver(false);
+        final ArchivingMediaDriver archivingMediaDriver = launchArchivingMediaDriver(true);
         final AeronArchive aeronArchive = connect();
         final AtomicBoolean running = new AtomicBoolean(true);
         final AtomicReference<Throwable> error = new AtomicReference<>();
-        final CountDownLatch publisherStarted = new CountDownLatch(1);
+        final CountDownLatch nodeStarted = new CountDownLatch(1);
 
-        final Thread archiveNode = new Thread(
+        final Thread echoNode = new Thread(
             () ->
             {
-                publisherStarted.countDown();
+                nodeStarted.countDown();
 
-                try (ArchiveNode node =
-                    new ArchiveNode(running, archivingMediaDriver, aeronArchive, false))
+                try (EchoNode node =
+                    new EchoNode(running, archivingMediaDriver.mediaDriver(), aeronArchive.context().aeron(), false))
                 {
                     node.run();
                 }
@@ -65,18 +65,18 @@ class LiveReplayMessageTransceiverTest
                     error.set(t);
                 }
             });
-        archiveNode.setName("archive-node");
-        archiveNode.setDaemon(true);
-        archiveNode.start();
+        echoNode.setName("echo-node");
+        echoNode.setDaemon(true);
+        echoNode.start();
 
         final LongArrayList timestamps = new LongArrayList(messages, MIN_VALUE);
-        final LiveReplayMessageTransceiver messageTransceiver = new LiveReplayMessageTransceiver(
-            archivingMediaDriver.mediaDriver(),
+        final LiveRecordingMessageTransceiver messageTransceiver = new LiveRecordingMessageTransceiver(
+            archivingMediaDriver,
             aeronArchive,
-            false,
+            true,
             timestamp -> timestamps.addLong(timestamp));
 
-        publisherStarted.await();
+        nodeStarted.await();
 
         messageTransceiver.init(configuration);
         try
@@ -105,13 +105,10 @@ class LiveReplayMessageTransceiverTest
         finally
         {
             running.set(false);
-            archiveNode.join();
+            echoNode.join();
             messageTransceiver.destroy();
-            CloseHelper.closeAll(aeronArchive, archivingMediaDriver);
-            archivingMediaDriver.mediaDriver().context().deleteAeronDirectory();
-            archivingMediaDriver.archive().context().deleteDirectory();
         }
 
-        assertArrayEquals(LongStream.range(1_000, 1_000 + messages).toArray(), timestamps.toLongArray());
+        assertArrayEquals(range(1_000, 1_000 + messages).toArray(), timestamps.toLongArray());
     }
 }
