+++ src/test/java/uk/co/real_logic/benchmarks/rtt/aeron/PingPongTest.java
@@ -15,51 +15,63 @@
  */
 package uk.co.real_logic.benchmarks.rtt.aeron;
 
-import io.aeron.archive.ArchivingMediaDriver;
-import io.aeron.archive.client.AeronArchive;
+import io.aeron.Aeron;
 import io.aeron.driver.MediaDriver;
 import org.agrona.LangUtil;
 import org.agrona.collections.LongArrayList;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import uk.co.real_logic.benchmarks.rtt.Configuration;
 
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
+import java.util.stream.LongStream;
 
-import static io.aeron.archive.client.AeronArchive.connect;
-import static java.lang.Long.MIN_VALUE;
-import static java.util.stream.LongStream.range;
+import static io.aeron.Aeron.connect;
+import static java.lang.System.clearProperty;
+import static java.lang.System.setProperty;
+import static org.agrona.CloseHelper.closeAll;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
-import static uk.co.real_logic.benchmarks.rtt.aeron.AeronUtil.launchArchivingMediaDriver;
+import static uk.co.real_logic.benchmarks.rtt.aeron.AeronUtil.EMBEDDED_MEDIA_DRIVER_PROP_NAME;
+import static uk.co.real_logic.benchmarks.rtt.aeron.AeronUtil.launchEmbeddedMediaDriverIfConfigured;
 
-class LiveRecordingMessageTransceiverTest
+class PingPongTest
 {
+    @BeforeAll
+    static void before()
+    {
+        setProperty(EMBEDDED_MEDIA_DRIVER_PROP_NAME, "true");
+    }
+
+    @AfterAll
+    static void after()
+    {
+        clearProperty(EMBEDDED_MEDIA_DRIVER_PROP_NAME);
+    }
+
     @Test
     void test() throws Exception
     {
-
         final int messages = 1_000_000;
         final Configuration configuration = new Configuration.Builder()
             .numberOfMessages(messages)
-            .messageTransceiverClass(LiveRecordingMessageTransceiver.class)
+            .messageTransceiverClass(PlainMessageTransceiver.class)
             .build();
 
-        final ArchivingMediaDriver archivingMediaDriver = launchArchivingMediaDriver(true);
-        final MediaDriver mediaDriver = archivingMediaDriver.mediaDriver();
-        final AeronArchive aeronArchive = connect();
+        final MediaDriver mediaDriver = launchEmbeddedMediaDriverIfConfigured();
+        final Aeron aeron = connect();
         final AtomicBoolean running = new AtomicBoolean(true);
         final AtomicReference<Throwable> error = new AtomicReference<>();
         final CountDownLatch publisherStarted = new CountDownLatch(1);
-
 
         final Thread echoNode = new Thread(
             () ->
             {
                 publisherStarted.countDown();
 
-                try (EchoNode node =
-                    new EchoNode(running, mediaDriver, aeronArchive.context().aeron(), false))
+                try (EchoNode node = new EchoNode(running, mediaDriver, aeron, false))
                 {
                     node.run();
                 }
@@ -72,11 +84,11 @@ class LiveRecordingMessageTransceiverTest
         echoNode.setDaemon(true);
         echoNode.start();
 
-        final LongArrayList timestamps = new LongArrayList(messages, MIN_VALUE);
-        final LiveRecordingMessageTransceiver messageTransceiver = new LiveRecordingMessageTransceiver(
-            archivingMediaDriver,
-            aeronArchive,
-            true,
+        final LongArrayList timestamps = new LongArrayList(messages, Long.MIN_VALUE);
+        final PlainMessageTransceiver messageTransceiver = new PlainMessageTransceiver(
+            mediaDriver,
+            aeron,
+            false,
             timestamp -> timestamps.addLong(timestamp));
 
         publisherStarted.await();
@@ -110,8 +122,14 @@ class LiveRecordingMessageTransceiverTest
             running.set(false);
             echoNode.join();
             messageTransceiver.destroy();
+            closeAll(aeron, mediaDriver);
+            mediaDriver.context().deleteAeronDirectory();
         }
 
-        assertArrayEquals(range(1_000, 1_000 + messages).toArray(), timestamps.toLongArray());
+        if (null != error.get())
+        {
+            LangUtil.rethrowUnchecked(error.get());
+        }
+        assertArrayEquals(LongStream.range(1_000, 1_000 + messages).toArray(), timestamps.toLongArray());
     }
 }
